+++ src/WildcardPattern.php
@@ -27,6 +27,7 @@ final class WildcardPattern
     private bool $matchSlashesExactly = false;
     private bool $matchLeadingPeriodExactly = false;
     private bool $ignoreCase = false;
+    private bool $matchEnding = false;
     private string $pattern;
 
     /**
@@ -77,7 +78,7 @@ final class WildcardPattern
         }
 
         $pattern = strtr(preg_quote($pattern, '#'), $replacements);
-        $pattern = '#^' . $pattern . '$#us';
+        $pattern = '#' . ($this->matchEnding ? '' : '^') . $pattern . '$#us';
 
         if ($this->ignoreCase) {
             $pattern .= 'i';
@@ -133,6 +134,19 @@ final class WildcardPattern
     {
         $new = clone $this;
         $new->matchLeadingPeriodExactly = $flag;
+        return $new;
+    }
+
+    /**
+     * Match ending only.
+     * By default wildcard pattern matches string exactly. By using this mode, beginning of the string could be anything.
+     * @param bool $flag
+     * @return self
+     */
+    public function withEnding(bool $flag = true): self
+    {
+        $new = clone $this;
+        $new->matchEnding = $flag;
         return $new;
     }
 }